+++ core/dbt/config/profile.py
@@ -52,7 +52,15 @@ def read_profile(profiles_dir: str) -> Dict[str, Any]:
     if os.path.isfile(path):
         try:
             contents = load_file_contents(path, strip=False)
-            return load_yaml_text(contents)
+            yaml_content = load_yaml_text(contents)
+            if not yaml_content:
+                msg = f'The profiles.yml file at {path} is empty'
+                raise DbtProfileError(
+                    INVALID_PROFILE_MESSAGE.format(
+                        error_string=msg
+                    )
+                )
+            return yaml_content
         except ValidationException as e:
             msg = INVALID_PROFILE_MESSAGE.format(error_string=e)
             raise ValidationException(msg) from e
@@ -280,7 +288,6 @@ class Profile(HasCredentials):
         # user_cfg is not rendered.
         if user_cfg is None:
             user_cfg = raw_profile.get('config')
-
         # TODO: should it be, and the values coerced to bool?
         target_name, profile_data = cls.render_profile(
             raw_profile, profile_name, target_override, renderer
@@ -335,7 +342,15 @@ class Profile(HasCredentials):
         # First, we've already got our final decision on profile name, and we
         # don't render keys, so we can pluck that out
         raw_profile = raw_profiles[profile_name]
-
+        if not raw_profile:
+            msg = (
+                f'Profile {profile_name} in profiles.yml is empty'
+            )
+            raise DbtProfileError(
+                INVALID_PROFILE_MESSAGE.format(
+                    error_string=msg
+                )
+            )
         user_cfg = raw_profiles.get('config')
 
         return cls.from_raw_profile_info(
@@ -373,7 +388,6 @@ class Profile(HasCredentials):
         raw_profiles = read_profile(args.profiles_dir)
         profile_name = cls.pick_profile_name(getattr(args, 'profile', None),
                                              project_profile_name)
-
         return cls.from_raw_profiles(
             raw_profiles=raw_profiles,
             profile_name=profile_name,